Fix type error by copying over "typeOf" function
At some point after llvm-hs-pure 9.0.0 branch, a commit was introduced
that changed `typeOf` to have a different type signature/implementation.
It was changed from a (partial) pure function to a total monadic
function, that requires a MonadModuleBuilder constraint.

While this is a good change, it is incompatible with the pretty printer
library that is used in llvm-hs-pretty which only works in a pure
context.

For now, the old `typeOf` function is copied over, but a better fix
would be to use the new version of `typeOf`. This will probably require
writing a new pretty printer library from the ground up, just to be able
to deal with monadic effects.